Minmax works
Debugged for a long time a situation when on debuging Intel projectors
and backprojectors segfaults on calling them. It might be related to the
fact that the last parameter of the kernel is float and this parameter
is on 12 position. Adding last cl_int2 dummy parameter seems to fix the
problem. Problem seems to be related to
https://community.amd.com/t5/opencl/opencl-cpu-number-of-kernel-arguments/td-p/43068
but was not able to find a reason.

Fixed minmax backprojector by initializing volume by infinity.

Into kniha add alg function to call FLOATcutting_voxel_minmaxbackproject
and fixed that call by dummy factor. Will test it and consider doing so
for other functions that segfaults.