FIX zero division in computing polygons
The problem was in the function include.cl:exactIntersectionPolygons0

Was triggered when one of
    REAL FproductVD1NEG = -vd1 * Fvector.x;
    REAL FproductVD3NEG = -vd3 * Fvector.y;
were zero.

This was causing zero division problems.

When FproductVD1NEG or FproductVD3NEG are zero then basically all the segment is projected to
given index but we need to achieve p \in [0,1] and q\in[0,1] so that we have to solve it individually
I put it int the way that now function include.cl:exactIntersectionPolygons0 greedy selects the biggest area possible.