Major code remake for local NDRange parsing
OpenCl kernel needs to supply local NDRange controlling partitioning
In projectors, especially barrier implementation this setting might have
tremendous impact on performance.
On each device there is maximum admissible size of workers in one
working group. So this parameter needs to be checked againts it.
Projector and reconstructor might use the same setting but it was
previously in each class.
New classes in NDRange folder to provide this logic.
Setting reasonable defaults for kct-pb and kct-pb2d
CBCT is yet to be fully change.