kct-pb2d-backprojector --backprojector-natural-scaling
Especially for FBP we do not have backprojector as simple AT operator
but there is an additional scaling factor by pi/(anglecount*pdimx) for
convenience there is a switch to apply this factor to kct-pb2d-backprojector
For this logic to work there is additional scaling factor parameter,
wchich defaults to 1.0f added to BasePBCTOperator and BasePBCT2DOperator
project and backproject methods. CArmArguments was updated and special
function addBackprojectorArgs() added to insert switch for this
functionality.