Rename of identifiers containing Sidon to Siddon
The name Sidon was misspelled in several places throughout the project.
Renamed all instances of "Sidon" to "Siddon" to accurately reflect the real name of Robert Siddon, the author of the Siddon projector, out of respect for his contribution.

Rename of sidon to siddon

There was misspelling of many names through this project using sidon or
Siddon
Renamed instances of "Sidon" to "Siddon" to reflect the real name of Robert Siddon, the author of the Siddon projector idea, out of respect for his contribution.