FIX: implementation of PartialPBCT2DOperator
The implementation for PBCT2D was following the implementation of PBCT
Projection matrix was actually used to properly offset along z axis and
to get relation between projection and volume. This is however not
working for current PBCT2D model, where we avoid using any kind of Z
axis and use slice to slice correspondence. With the propor offseting of
projections is however possible to achieve projection of different x and
b slices, k_from and k_to is to control slab width and position.